// client/src/types/exercise.ts
export type exerciseType =
  | "Barbell"
  | "Dumbbell"
  | "Machine"
  | "Cable"
  | "Weighted Bodyweight"
  | "Assisted Bodyweight"
  | "Cardio"
  | "Other";

export const exerciseTypeList = [
  "Barbell",
  "Dumbbell",
  "Machine",
  "Cable",
  "Weighted Bodyweight",
  "Assisted Bodyweight",
  "Cardio",
  "Other",
];

export type exerciseMuscleGroup =
  | "None"
  | "Abs"
  | "Arms"
  | "Back"
  | "Chest"
  | "Legs"
  | "Shoulders"
  | "Full Body"
  | "Cardio"
  | "Other";

export const exerciseMuscleGroupList = [
  "None",
  "Abs",
  "Arms",
  "Back",
  "Chest",
  "Legs",
  "Shoulders",
  "Full Body",
  "Cardio",
  "Other",
];